creates util for gurbani
