+++ src/js/util/dom/index.ts
@@ -1,0 +1,2 @@
+export * from './copy-to-clipboard';
+export * from './make-selection';
