+++ src/js/util/dom/make-selection.ts
@@ -1,0 +1,7 @@
+
+export const makeSelection = (selectedDiv: HTMLElement) => {
+  window.getSelection().removeAllRanges();
+  var range = document.createRange();
+  range.selectNode(selectedDiv);
+  window.getSelection().addRange(range);
+};
