+++ src/js/util/gurbani/index.ts
@@ -1,0 +1,3 @@
+export * from './verses-to-gurbani';
+export * from './get-highlighting-endpoints';
+export * from './get-highlight-indices';
